feat: add wasm support
