+++ pkg/app/server/render/html.go
@@ -43,13 +43,8 @@ package render
 
 import (
 	"html/template"
-	"log"
-	"sync"
-	"time"
 
-	"github.com/cloudwego/hertz/pkg/common/hlog"
 	"github.com/cloudwego/hertz/pkg/protocol"
-	"github.com/fsnotify/fsnotify"
 )
 
 // Delims represents a set of Left and Right delimiters for HTML template rendering.
@@ -107,101 +102,4 @@ func (r HTML) Render(resp *protocol.Response) error {
 // WriteContentType (HTML) writes HTML ContentType.
 func (r HTML) WriteContentType(resp *protocol.Response) {
 	writeContentType(resp, htmlContentType)
-}
-
-type HTMLDebug struct {
-	sync.Once
-	Template        *template.Template
-	RefreshInterval time.Duration
-
-	Files   []string
-	FuncMap template.FuncMap
-	Delims  Delims
-
-	reloadCh chan struct{}
-	watcher  *fsnotify.Watcher
-}
-
-func (h *HTMLDebug) Instance(name string, data interface{}) Render {
-	h.Do(func() {
-		h.startChecker()
-	})
-
-	select {
-	case <-h.reloadCh:
-		h.reload()
-	default:
-	}
-
-	return HTML{
-		Template: h.Template,
-		Name:     name,
-		Data:     data,
-	}
-}
-
-func (h *HTMLDebug) Close() error {
-	if h.watcher == nil {
-		return nil
-	}
-	return h.watcher.Close()
-}
-
-func (h *HTMLDebug) reload() {
-	h.Template = template.Must(template.New("").
-		Delims(h.Delims.Left, h.Delims.Right).
-		Funcs(h.FuncMap).
-		ParseFiles(h.Files...))
-}
-
-func (h *HTMLDebug) startChecker() {
-	h.reloadCh = make(chan struct{})
-
-	if h.RefreshInterval > 0 {
-		go func() {
-			hlog.SystemLogger().Debugf("[HTMLDebug] HTML template reloader started with interval %v", h.RefreshInterval)
-			for range time.Tick(h.RefreshInterval) {
-				hlog.SystemLogger().Debugf("[HTMLDebug] triggering HTML template reloader")
-				h.reloadCh <- struct{}{}
-				hlog.SystemLogger().Debugf("[HTMLDebug] HTML template has been reloaded, next reload in %v", h.RefreshInterval)
-			}
-		}()
-		return
-	}
-
-	watcher, err := fsnotify.NewWatcher()
-	if err != nil {
-		log.Fatal(err)
-	}
-	h.watcher = watcher
-	for _, f := range h.Files {
-		err := watcher.Add(f)
-		hlog.SystemLogger().Debugf("[HTMLDebug] watching file: %s", f)
-		if err != nil {
-			hlog.SystemLogger().Errorf("[HTMLDebug] add watching file: %s, error happened: %v", f, err)
-		}
-
-	}
-
-	go func() {
-		hlog.SystemLogger().Debugf("[HTMLDebug] HTML template reloader started with file watcher")
-		for {
-			select {
-			case event, ok := <-watcher.Events:
-				if !ok {
-					return
-				}
-				if event.Op&fsnotify.Write == fsnotify.Write {
-					hlog.SystemLogger().Debugf("[HTMLDebug] modified file: %s, html render template will be reloaded at the next rendering", event.Name)
-					h.reloadCh <- struct{}{}
-					hlog.SystemLogger().Debugf("[HTMLDebug] HTML template has been reloaded")
-				}
-			case err, ok := <-watcher.Errors:
-				if !ok {
-					return
-				}
-				hlog.SystemLogger().Errorf("error happened when watching the rendering files: %v", err)
-			}
-		}
-	}()
 }
